backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for API calls
httpx==0.25.2
aiohttp==3.9.1

# Google Gemini AI
google-generativeai==0.3.1

# ElevenLabs TTS (if official SDK exists, otherwise use httpx)
elevenlabs==0.2.27

# Environment variables
python-dotenv==1.0.0

# CORS
python-jose[cryptography]==3.3.0

# Sentiment analysis for emotion mismatch
vaderSentiment==3.3.2

